+++ static/routes.config.php
@@ -17,7 +17,7 @@ return [
 
 	'/.well-known' => [
 		'/host-meta'      => [Module\WellKnown\HostMeta::class,     [R::GET]],
-		'/nodeinfo'       => [Module\NodeInfo::class,               [R::GET]],
+		'/nodeinfo'       => [Module\WellKnown\NodeInfo::class,     [R::GET]],
 		'/webfinger'      => [Module\Xrd::class,                    [R::GET]],
 		'/x-social-relay' => [Module\WellKnown\XSocialRelay::class, [R::GET]],
 	],
